# demo/swagger/valencia/channel_message.yaml
...
swagger: "2.0"
...
paths:
...
  /messages/download/csv:
    get:
      x-swagger-router-controller: message
      description: メッセージー一覧をcsvファイルでダウンロード
      operationId: message#download
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt:
        - api:access
      summary: メッセージー一覧ダウンロード
      x-submit-label: メッセージー一覧をダウンロードする
      tags:
      - message
...